pyproject.toml
[tool.pdm]

[project]
name = "NetworkMonitor"
version = "0.0.1"
description = "Simple network monitor for my small server"
authors = [
    {name = "jaypyles", email = "jpylesbusiness@gmail.com"},
]
dependencies = [
    "speedtest-cli>=2.1.3",
    "Flask[async]>=2.3.2",
    "asyncio>=3.4.3",
]
requires-python = ">=3.11"
license = {text = "MIT"}


[tool.isort]
length_sort = "1"
profile = "black"
sections = "STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
import_heading_stdlib = "STL"
import_heading_thirdparty = "PDM"
import_heading_firstparty = "LOCAL"
import_heading_localfolder = "LOCAL"

[tool.pyright]
include = ["network_usage"]
exclude = ["**/node_modules", "**/__pycache__"]
ignore = []
defineConstant = { DEBUG = true }
stubPath = ""

reportMissingImports = true
reportMissingTypeStubs = false

pythonVersion = "3.9"
pythonPlatform = "Linux"
